update test for pression convertion fixedpoint

+++ src/software/firmware/includes/pression.h
@@ -21,7 +21,7 @@
  * @param sensorValue 
  * @return the pressure in mmH2O 
  */
-int convertSensor2Pressure(uint16_t sensorValue);
+int16_t convertSensor2Pressure(uint16_t sensorValue);
 
 /**
  * Get the measured or simulated pressure for the feedback control (in mmH2O)
